requirements.txt
Flask>=0.12
Flask-cors>=3.0
newspaper3k>=0.1
